+++ user-api/src/main/java/com/biengual/userapi/content/presentation/ContentPublicController.java
@@ -1,10 +1,8 @@
 package com.biengual.userapi.content.presentation;
 
 import com.biengual.userapi.content.application.ContentFacade;
-import com.biengual.userapi.content.application.ContentService;
+import com.biengual.userapi.content.domain.ContentService;
 import com.biengual.userapi.content.domain.ContentInfo;
-import com.biengual.userapi.content.domain.dto.ContentRequestDto;
-import com.biengual.userapi.content.domain.dto.ContentResponseDto;
 import com.biengual.userapi.content.domain.enums.ContentType;
 import com.biengual.userapi.message.ResponseEntityFactory;
 import com.biengual.userapi.swagger.content.SwaggerContentByScrapCount;
