+++ user-api/src/main/java/com/biengual/userapi/swagger/content/SwaggerContentPreview.java
@@ -3,7 +3,7 @@ package com.biengual.userapi.swagger.content;
 import java.util.List;
 import java.util.Map;
 
-import com.biengual.userapi.content.domain.dto.ContentResponseDto;
+import com.biengual.userapi.content.presentation.ContentResponseDto;
 import com.biengual.userapi.swagger.SwaggerReturnInterface;
 
 public class SwaggerContentPreview
