+++ user-api/src/main/java/com/biengual/userapi/swagger/content/SwaggerContentUpdate.java
@@ -1,6 +1,6 @@
 package com.biengual.userapi.swagger.content;
 
-import com.biengual.userapi.content.domain.dto.ContentResponseDto;
+import com.biengual.userapi.content.presentation.ContentResponseDto;
 import com.biengual.userapi.swagger.SwaggerReturnInterface;
 
 public class SwaggerContentUpdate extends SwaggerReturnInterface<ContentResponseDto.ContentUpdateResponseDto> {
